Set username and home directory from environment variables
Use environment variables $USER and $HOME for home.username and
home.homeDirectory respectfully. This requires passing the "--impure"
option to "home-manager switch" the first time it's run after the
change. This allows the use of this configuration on other machines
without having to add per-host configurations to the repository. I'm
okay with this for now.